Added some cross-referencing of objects:   * Markets <-> Commodities   * Facilities -> Commodities   * Regions -> Allowed facilities
Can now define a set of markets with commodities.
Can now define a set of available facilities with in/out commodities.
Can now define a set of regions with facilities.

Beginning to fight against input parsing.
Commodities can only be fully defined as a paramter of a market.
However, care is taken to only define new commodities if they have not
already been defined.
Commodities can be reused in facilities.