Changed Model ID's to have a single, global ID-space.  This change required me to make the following adjustments in the code:
    * changed ModelList to be a map instead of a vector to accomodate 
      non-consecutive ID's.

    * added begin(ModelType) and end(ModelType) methods to the Logician in 
      order to preserve iterating functionality over models for classes 
      external to logician (currently only used by bookkeeper).

    * Added next_id_ static member var to Model class.  Default model 
      constructor automatically sets the ID_ var and increments next_id_. 
      Removed nextID vars and nextID++ statements from FacilityMode, 
      MarketModel, etc. constructors.

I also deleted some comments that I felt were not helpful at all.  We 
should put some effort into deleting and/or improving bad comments. 
Examples of unhelpful comments:

ConverterModel.h
/**
 * The ConverterModel class is the abstract class/interconve used by all
 * converter models
 * 
*/
Yes, we know it is a class.  And we probably could guess that more 
specific converter types inherit from this.

Material.h
/**
 * type definition for isotopes
 */
What is the variable used for? How is it used.  This isn't worthless, 
but it's not great.