Removed svn:executable property from files that probably should not be executable. This is mostly aestheic, in my opinion, but could be perceived as a security issue, I imagine.
Anyway, files that should not be executable inclue *.cpp, *.h, *.txt, *.xml, 
and others. It may be worth checking that your default file permissions 
do not include the executable option. The way to do this on a unix/linux
system is by setting umask in your .bashrc file to something like -rw-r--r-- 
(umask 002) or -rw-rw---- (umask 022). The default on the CAE machines seems to be 
022, which is fine. 

I know my settings were wrong for a while, so I'm at fault for many of these 
files having been executables early on, but it's also the case that I've seen 
more files popping up recently with executable permissions, so I thought I'd 
wipe it clean and mention it here in the hopes that we can keep it clean going 
forward.