Changed construct and destruct function symbols according to the suggestion (by Steve) referred to in issue89 . This commit closes issue89.
This change was necessary in order to demonstrate that two concrete 
FacilityModels can tested with the same FacilityModelTests interface, 
toward the completion of issue118 and issue105 .
This demonstration is shown with the SourceFac.FacilityModelTests and the 
SinkFac.FacilityModelTests that run using googletests parameterized test 
fixture capability. This commit provides a proof of principle for the Abstract 
Test Pattern which will be used throughout the code for the testing of 
abstract interfaces. 

The segfault reported in r566 was diagnosed and corrected in the 
FacilityModelTests copyFacility test.