This commit adds a static output directory string for each class including TimeAgent, RegionModel, InstModel, FacilityModel, and their respective inheritency children. The string itself if private, but there is a public getter function.
Additionally, each string is built upon its inheritence. For example, The source facility has a privte string called outputDir_. One accesses the full output string for source facility by SourceFacility::outputDir(). The return string for this function is actually FacilityMode::outputDir.append(outputDir_).

The purpose of this methodology is to allow for there to be a bookkeeper function which takes arguments of outputDir and parameterList and adds these to a container for eventual writing to the output database. Accordingly, each atomic entity can decide when to ask the bookkeeper to record something (such as a state change) by prompting: BI->record(ThisClass::outputDir, aParameterList).