Adds Stub*Input.xml files, contributing to issue #24.
The way I have chosen to do this is quite redundant.

However, putting these files in the folders with the models decreases
the burden on the developer in comparison with the option of adding
just one StubInput.xml file in the core/input/ directory.

I couldn't decide which of these options was best, so I just picked
this one. Any thoughts? Is redundancy preferable to an extra step in
development, or is it preferable to keep the input files out of the src
directory?