staging for final isovector work. testing of recipelogger complete.
changed to using shared pointers for compositions and compmaps. recipelogger appears to be done, so testing and documentation need to be written. then on to adjusting the much simplified isovector.

updated make process to write and run new recipe tests

wrote tests for isotopic definitions... still have a lot of thinking to do about this topic..

adding some tests for Isotopic Definitions. name will be changed soon.