Add internal tag to cycpp. Modify derived_init tag behavior:
'internal' tag now makes state variables invisible in the xml schema, but
still persisted/loaded normally.  'internal' vars can have default values.
'derived_init' vars now show up in the xml schema as either required or
optional depending on whether a default is specified and the derived_init
value always run after initializing to a default and/or from a user specified
value for the variable.  If a var is both internal and derived_init, then it
is not in the schema but can still have a default value specified.