Preference adjustment has been updated to include the managers of both requesters and bidders. This will let institution and region interactions be based either trader entity. Only bidders now have no influence on the preference adjustment process, which could be updated in the future if needed.
There is a notable deprecation of the *Agent* AdjustPrefs API. As was
true previously, only *managers* of trading agents are called through
this API, i.e., institutions and regions in the RIF model. The new
AdjustPref API includes the request, bid, preference, and whether the
agent is a manager of the requester or bidder (called
TradeSense). Each trade has its preference adjusted
individually. Anything more complex will require much more difficult
logic in the DRE *and* in the agent API.