fixed agent memory leak:
Previously, when a context was destructed, it would only delete the
root-agents (agents without parents).  Later on, in order to enable agents to
deploy other agents and then leave the simulation, we changed agents to not
destruct their children when they are destructed, but the context was not
updated to destruct non-root agents too.  This fixes that.