Deprecate `ProgTranslator::Context` without changing API/ABI.
This addresses #1279 by returning to the situation where there is
no leakage of COIN headers and thus no need for any archetypes that
use the `building_manager` to know about COIN headers.

The previous version of this caused an API/ABI change that, although
unlikely to be a problem for anyone, is contrary to our stability
philosophy.  This version maintains the API/ABI stability by

* changing all references (2 in `prog_translator.h` & 3 in `building_manager.cc`)
  of `ProgTranslator::Context` to refer to `ProgTranslatorContext`
* reintroducing the `ProgTranslator::Context` class
* explicitly defining its constructor and destructor
* implementing said con/destructors to throw a DepricatedApiError
* adding the DepricatedApiError
* adding a test to demonstrate that this error is thrown